Move display shortcuts implementation to a new file so we can use React.
This also shows the shortcuts in a simple dialog. The styles are hardcoded at this point.
